More import fixes and refactoring.
