+++ lib/commands/common.ts
@@ -1,4 +1,4 @@
-import { C8yPact } from "../../shared/c8ypact";
+import { C8yPact } from "@shared/c8ypact";
 
 export {};
