+++ lib/commands/intercept.ts
@@ -1,5 +1,4 @@
-import { C8yDefaultPact } from "../../shared/c8ypact";
-import { C8yPactUrlMatcher } from "../../shared/c8ypact/urlmatcher";
+import { C8yDefaultPact, C8yPactUrlMatcher } from "@shared/c8ypact";
 import {
   HTTP_METHODS,
   STATIC_RESPONSE_KEYS,
