+++ lib/pacts/runner.ts
@@ -1,4 +1,4 @@
-import { C8yPact, C8yPactInfo, C8yPactRecord } from "../../shared/c8ypact";
+import { C8yPact, C8yPactInfo, C8yPactRecord } from "@shared/c8ypact";
 
 const { _ } = Cypress;
